requirements.txt
flask
flask-wtf
flask_login
flask_sqlalchemy
gunicorn
python-decouple
sqlalchemy
gunicorn
uuid
